+++ src/modules/metrics/metrics.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { MetricsService } from './metrics.service';
+import { MetricsInterceptor } from './metrics.interceptor';
+import { APP_INTERCEPTOR } from '@nestjs/core';
+import { MetricsController } from './metrics.controller';
+
+@Module({
+  controllers: [MetricsController],
+  providers: [MetricsService, { provide: APP_INTERCEPTOR, useClass: MetricsInterceptor }],
+  exports: [MetricsService],
+})
+export class MetricsModule {}